[MOV] portal: split mail and chatter controllers, like in mail
Put portal thread-like controllers in portal_thread file, like classic
thread code is split from other mail controllers. Mail file contains
controllers about links received in emails (like 'View Documents' or
message redirection).

Prepares Task-4430903: [mail] Stop aufollow madness !
Prepares Task-4273479: [mail] Email-like recipients
