+++ addons/portal/controllers/__init__.py
@@ -1,8 +1,7 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from . import attachment
 from . import mail
 from . import message_reaction
 from . import portal
+from . import portal_thread
 from . import thread
 from . import web
