+++ addons/portal_rating/controllers/portal_chatter.py
@@ -1,11 +1,9 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo.osv import expression
 
-from odoo.addons.portal.controllers import mail
+from odoo.addons.portal.controllers import portal_thread
 
 
-class PortalChatter(mail.PortalChatter):
+class PortalChatter(portal_thread.PortalChatter):
 
     def _setup_portal_message_fetch_extra_domain(self, data):
         domains = [super()._setup_portal_message_fetch_extra_domain(data)]
