+++ addons/website_slides/controllers/mail.py
@@ -4,7 +4,7 @@ from werkzeug.exceptions import NotFound, Forbidden
 
 from odoo import http
 from odoo.http import request
-from odoo.addons.portal.controllers.mail import PortalChatter
+from odoo.addons.portal.controllers.portal_thread import PortalChatter
 from odoo.tools import plaintext2html, html2plaintext
